Move HKDevice into connection
