+++ homeassistant/components/homekit_controller/const.py
@@ -4,6 +4,9 @@ DOMAIN = 'homekit_controller'
 KNOWN_DEVICES = "{}-devices".format(DOMAIN)
 CONTROLLER = "{}-controller".format(DOMAIN)
 
+HOMEKIT_DIR = '.homekit'
+PAIRING_FILE = 'pairing.json'
+
 # Mapping from Homekit type to component.
 HOMEKIT_ACCESSORY_DISPATCH = {
     'lightbulb': 'light',
